Fixed vulnerability with arbitrary argument specification
Mailicous user could specify post $url to be something like
"--malicious-option=/etc/shadow" in order to make youtube-dl
behave in unexpected way. Youtube-dl *should* exit because
of missing URL, but I'm not sure of it and I'm not risking it.

This fix uses `--` (see
https://soptik.tech/articles/beware-of-shell-globs.html)
which should mean that $url will *always* be interpreted
as filename rather than a youtube-dl option.